requirements.txt
finnhub-python==2.4.19
python-dotenv==1.0.0
requests==2.31.0
openai==2.5.0
